fix(base): narrow workflow validation scope
1. Keep preflight checks limited to DeleteRecordTrigger and LarkMessageAction contracts
2. Move shared validation and E2E coverage into existing workflow files
3. Restore unrelated update and coverage behavior to the pre-PR baseline

```ai-signature
改动范围: shortcuts/base 的 workflow create/update 校验、既有测试文件与 lark-base workflow 指南
思考过程: 保留 014 和 026 的发送前防线，删除顶层 schema 校验和重复解析，并把原新增文件内容合并到现有文件以缩小有效 diff
改动原因: PR 仅需修复不支持的删除触发器与消息动作缺字段问题，现有实现扩大了契约和文件数量
Break Change: 否
```

AI-SHA256: 55da68b4a04ec4c0250365527417936252760795db3044e7a68e2664d33c606b
