+++ shortcuts/base/base_shortcuts_test.go
@@ -627,7 +627,7 @@ func TestBaseDashboardHelpGuidesAgents(t *testing.T) {
 }
 
 func TestBaseWorkflowHelpGuidesAgents(t *testing.T) {
-	if got, want := BaseWorkflowUpdate.Description, "Replace a workflow's full definition (title and steps) in a base"; got != want {
+	if got, want := BaseWorkflowUpdate.Description, "Replace a workflow's full definition (title and/or steps) in a base"; got != want {
 		t.Fatalf("workflow update description=%q, want %q", got, want)
 	}
 
@@ -674,12 +674,8 @@ func TestBaseWorkflowHelpGuidesAgents(t *testing.T) {
 			wantTips: []string{
 				"lark-cli base +workflow-update --base-token <base_token> --workflow-id <workflow_id> --json @workflow.json",
 				"PUT uses full replacement semantics",
-				"title and steps are both required",
-				"explicitly send \"steps\":[]",
-				"do not omit steps",
 				"Use +workflow-get first",
 				"returned title and steps",
-				"empty LarkMessageAction data.btn_list",
 				"workflow-id must start with wkf",
 				"Updating does not enable or disable",
 				"do not invent steps[].type/data/next/children from natural language",
