+++ shortcuts/base/workflow_update.go
@@ -7,13 +7,14 @@ import (
 	"context"
 	"strings"
 
+	"github.com/larksuite/cli/errs"
 	"github.com/larksuite/cli/shortcuts/common"
 )
 
 var BaseWorkflowUpdate = common.Shortcut{
 	Service:     "base",
 	Command:     "+workflow-update",
-	Description: "Replace a workflow's full definition (title and steps) in a base",
+	Description: "Replace a workflow's full definition (title and/or steps) in a base",
 	Risk:        "write",
 	Scopes:      []string{"base:workflow:update"},
 	AuthTypes:   []string{"user", "bot"},
@@ -24,9 +25,8 @@ var BaseWorkflowUpdate = common.Shortcut{
 	},
 	Tips: []string{
 		"lark-cli base +workflow-update --base-token <base_token> --workflow-id <workflow_id> --json @workflow.json",
-		"PUT uses full replacement semantics; title and steps are both required. To clear all steps, keep the current title and explicitly send \"steps\":[]; do not omit steps.",
+		"PUT uses full replacement semantics; omitting steps clears the existing workflow steps.",
 		"Use +workflow-get first, then build the update body from the returned title and steps; preserve every step field you do not intend to change.",
-		"The API may omit optional action fields such as an empty LarkMessageAction data.btn_list; keep those fields omitted unless you intend to set them.",
 		"workflow-id must start with wkf; do not pass a tbl table ID.",
 		"Step ids must be unique, and every next/children link must reference an existing step id.",
 		"Updating does not enable or disable a workflow; call +workflow-enable or +workflow-disable separately.",
@@ -44,7 +44,7 @@ var BaseWorkflowUpdate = common.Shortcut{
 		if err != nil {
 			return err
 		}
-		return validateWorkflowDefinition(body, workflowUpdateOperation)
+		return validateWorkflowDefinition(body)
 	},
 	DryRun: func(ctx context.Context, runtime *common.RuntimeContext) *common.DryRunAPI {
 		pc := newParseCtx(runtime)
@@ -73,4 +73,67 @@ var BaseWorkflowUpdate = common.Shortcut{
 		runtime.Out(data, nil)
 		return nil
 	},
+}
+
+// validateWorkflowDefinition enforces the workflow contracts shared by create
+// and update that the CLI can determine without replacing the server schema.
+func validateWorkflowDefinition(body map[string]interface{}) error {
+	steps, ok := body["steps"].([]interface{})
+	if !ok {
+		return nil
+	}
+
+	for index, rawStep := range steps {
+		step, ok := rawStep.(map[string]interface{})
+		if !ok {
+			continue
+		}
+		stepType, _ := step["type"].(string)
+		switch stepType {
+		case "DeleteRecordTrigger":
+			return errs.NewValidationError(
+				errs.SubtypeFailedPrecondition,
+				"--json.steps[%d].type %q is not supported: record deletion events are not available as Base workflow triggers",
+				index,
+				stepType,
+			).
+				WithParam("--json").
+				WithHint("No workflow request was sent. Keep the requested semantics unchanged; propose alternatives and write only after the user explicitly selects one.")
+		case "LarkMessageAction":
+			if err := validateWorkflowMessageAction(index, step["data"]); err != nil {
+				return err
+			}
+		}
+	}
+	return nil
+}
+
+func validateWorkflowMessageAction(index int, rawData interface{}) error {
+	data, ok := rawData.(map[string]interface{})
+	if !ok {
+		return workflowFieldError("--json.steps[%d].data for LarkMessageAction must be a JSON object", index)
+	}
+	if receiver, ok := data["receiver"].([]interface{}); !ok || len(receiver) == 0 {
+		return workflowFieldError("--json.steps[%d].data.receiver for LarkMessageAction must be a non-empty array", index)
+	}
+	if content, ok := data["content"].([]interface{}); !ok || len(content) == 0 {
+		return workflowFieldError("--json.steps[%d].data.content for LarkMessageAction must be a non-empty array", index)
+	}
+	if sendToEveryone, provided := data["send_to_everyone"]; provided {
+		if _, ok := sendToEveryone.(bool); !ok {
+			return workflowFieldError("--json.steps[%d].data.send_to_everyone for LarkMessageAction must be a boolean when provided", index)
+		}
+	}
+	if buttonList, provided := data["btn_list"]; provided {
+		if _, ok := buttonList.([]interface{}); !ok {
+			return workflowFieldError("--json.steps[%d].data.btn_list for LarkMessageAction must be an array when provided", index)
+		}
+	}
+	return nil
+}
+
+func workflowFieldError(format string, args ...any) error {
+	return errs.NewValidationError(errs.SubtypeInvalidArgument, format, args...).
+		WithParam("--json").
+		WithHint("Fix the reported field without inferring values or rewriting unrelated workflow data.")
 }
