+++ tests/cli_e2e/base/base_basic_workflow_test.go
@@ -68,3 +68,127 @@ func TestBase_BasicWorkflow(t *testing.T) {
 		assert.Equal(t, tableName, table.Get("name").String())
 	})
 }
+
+func TestBaseWorkflowUpdateDryRun(t *testing.T) {
+	result := runBaseDryRun(t, 0,
+		"base", "+workflow-update",
+		"--base-token", "app_x",
+		"--workflow-id", "wkf_x",
+		"--json", `{"title":"Reminder","steps":[{"type":"LarkMessageAction","data":{"receiver":[{"value_type":"user","value":{"id":"ou_x"}}],"content":[{"value_type":"text","value":"Review the request"}]}}]}`,
+	)
+
+	out := result.Stdout
+	require.Equal(t, "/open-apis/base/v3/bases/app_x/workflows/wkf_x", clie2e.DryRunGet(out, "api.0.url").String(), out)
+	require.Equal(t, "PUT", clie2e.DryRunGet(out, "api.0.method").String(), out)
+	require.Equal(t, "Reminder", clie2e.DryRunGet(out, "api.0.body.title").String(), out)
+	require.Equal(t, "LarkMessageAction", clie2e.DryRunGet(out, "api.0.body.steps.0.type").String(), out)
+	require.Equal(t, "ou_x", clie2e.DryRunGet(out, "api.0.body.steps.0.data.receiver.0.value.id").String(), out)
+	require.Equal(t, "Review the request", clie2e.DryRunGet(out, "api.0.body.steps.0.data.content.0.value").String(), out)
+	require.False(t, clie2e.DryRunGet(out, "api.0.body.steps.0.data.btn_list").Exists(), out)
+}
+
+func TestBaseWorkflowCreateDryRun(t *testing.T) {
+	result := runBaseDryRun(t, 0,
+		"base", "+workflow-create",
+		"--base-token", "app_x",
+		"--json", `{"title":"Reminder","client_token":"create_1","steps":[{"type":"LarkMessageAction","data":{"receiver":[{"value_type":"user","value":{"id":"ou_x"}}],"content":[{"value_type":"text","value":"Review the request"}]}}]}`,
+	)
+
+	out := result.Stdout
+	require.Equal(t, "/open-apis/base/v3/bases/app_x/workflows", clie2e.DryRunGet(out, "api.0.url").String(), out)
+	require.Equal(t, "POST", clie2e.DryRunGet(out, "api.0.method").String(), out)
+	require.Equal(t, "create_1", clie2e.DryRunGet(out, "api.0.body.client_token").String(), out)
+	require.Equal(t, "ou_x", clie2e.DryRunGet(out, "api.0.body.steps.0.data.receiver.0.value.id").String(), out)
+	require.Equal(t, "Review the request", clie2e.DryRunGet(out, "api.0.body.steps.0.data.content.0.value").String(), out)
+	require.False(t, clie2e.DryRunGet(out, "api.0.body.steps.0.data.btn_list").Exists(), out)
+}
+
+func TestBaseWorkflowDryRunRejectsInvalidDefinitions(t *testing.T) {
+	for _, tt := range []struct {
+		name        string
+		args        []string
+		wantSubtype string
+		wantPath    string
+	}{
+		{
+			name: "unsupported trigger",
+			args: []string{
+				"base", "+workflow-update",
+				"--base-token", "app_x",
+				"--workflow-id", "wkf_x",
+				"--json", `{"title":"Archive workflow","steps":[{"type":"DeleteRecordTrigger","data":{"table_id":"tbl_x"}}]}`,
+			},
+			wantSubtype: "failed_precondition",
+			wantPath:    "steps[0].type",
+		},
+		{
+			name: "missing message content",
+			args: []string{
+				"base", "+workflow-update",
+				"--base-token", "app_x",
+				"--workflow-id", "wkf_x",
+				"--json", `{"title":"Reminder","steps":[{"type":"LarkMessageAction","data":{"receiver":[{"value_type":"user","value":{"id":"ou_x"}}]}}]}`,
+			},
+			wantSubtype: "invalid_argument",
+			wantPath:    "content",
+		},
+		{
+			name: "invalid optional message field",
+			args: []string{
+				"base", "+workflow-create",
+				"--base-token", "app_x",
+				"--json", `{"title":"Reminder","client_token":"create_1","steps":[{"type":"LarkMessageAction","data":{"receiver":[{"value_type":"user","value":{"id":"ou_x"}}],"content":[{"value_type":"text","value":"Review the request"}],"send_to_everyone":"yes"}}]}`,
+			},
+			wantSubtype: "invalid_argument",
+			wantPath:    "send_to_everyone",
+		},
+	} {
+		t.Run(tt.name, func(t *testing.T) {
+			result := runBaseDryRun(t, 2, tt.args...)
+			require.Equal(t, "validation", gjson.Get(result.Stderr, "error.type").String(), result.Stderr)
+			require.Equal(t, tt.wantSubtype, gjson.Get(result.Stderr, "error.subtype").String(), result.Stderr)
+			require.Equal(t, "--json", gjson.Get(result.Stderr, "error.param").String(), result.Stderr)
+			require.Contains(t, gjson.Get(result.Stderr, "error.message").String(), tt.wantPath, result.Stderr)
+			require.Empty(t, result.Stdout)
+		})
+	}
+}
+
+func readWorkflowReference(t *testing.T, name string) string {
+	t.Helper()
+
+	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
+	t.Cleanup(cancel)
+
+	result, err := clie2e.RunCmd(ctx, clie2e.Request{
+		Args: []string{"skills", "read", "lark-base", "references/" + name},
+	})
+	require.NoError(t, err)
+	result.AssertExitCode(t, 0)
+	return result.Stdout
+}
+
+func TestWorkflowSkillDeliveryRegression(t *testing.T) {
+	guide := readWorkflowReference(t, "lark-base-workflow-guide.md")
+	for _, contract := range []string{
+		"创建/更新时重点构造 `title` 和 `steps`",
+		"`status` 通过 `+workflow-enable` 或 `+workflow-disable` 单独管理",
+	} {
+		require.Contains(t, guide, contract)
+	}
+	require.NotContains(t, guide, "创建/更新时重点构造 `title`、`status` 和 `steps`")
+}
+
+func TestWorkflowSchemaDeliveryRegression(t *testing.T) {
+	schema := readWorkflowReference(t, "lark-base-workflow-schema.md")
+	for _, contract := range []string{
+		"| `receiver` | 是 | 非空 ValueInfo[] |",
+		"| `content` | 是 | 非空 TextRefItem[] 消息内容 |",
+		"| `send_to_everyone` | 否 | boolean；",
+		"| `btn_list` | 否 | ButtonConfig[]；",
+		"`DeleteRecordTrigger`",
+		"明确选择该替代语义",
+	} {
+		require.Contains(t, schema, contract)
+	}
+}
